Backup storage for 2PC
The TwoPCBackend class can take any IBackend to serve as a simple
backup to it. The results of 2PC transactions are saved in the
backup, but only when the 2PC transaction commits! Queries are done
against it too, results merged.

The DTBackend now uses a SqlBackend as backup if it is so configured.
The trio of servers now keeps the DB up to date with all changes.